Add files via upload
Isebelle Fraser's diaries schema and two files encoded with it (wormwood and adams).
Info: I was preparing to turn everything in when the power went out in my house. It is now back on, and I believe that everything was saved. I looked over all the files to be sure, but if something seems off or didn’t get saved properly, please let me know and I will correct it ASAP!